Revert "Merge branch 'documentation_OREv2' into 'master'"
This reverts commit 01314f9f690d3ed655090dfc0ed1e0a4db8588da, reversing
changes made to 31c94c6f131b4b00b7367bc129ea7afcf22db8ad.

+++ src/main/java/fr/inra/oresing/rest/CsvRowValidationCheckResult.java
@@ -5,5 +5,5 @@ import lombok.Value;
 @Value
 public class CsvRowValidationCheckResult {
     ValidationCheckResult validationCheckResult;
-    long lineNumber;
+    int lineNumber;
 }
